Add kernels and representation for Image Buffers - Texture memory

+++ cmake-build-debug/kernels/kernel_const_data.cl
@@ -5,7 +5,7 @@ typedef struct
     unsigned int function;
     unsigned int maxInputs;
     unsigned int inputs[MAX_ARITY];
-    double inputsWeight[MAX_ARITY];
+    float inputsWeight[MAX_ARITY];
     int active;
 
 } Node;
@@ -15,8 +15,8 @@ typedef struct
     Node nodes[MAX_NODES];
     unsigned int output[MAX_OUTPUTS];
     unsigned int numActiveNodes;
-    double fitness;
-    double fitnessValidation;
+    float fitness;
+    float fitnessValidation;
 } Chromosome;
 
 typedef struct {
@@ -26,7 +26,7 @@ typedef struct {
 
 typedef struct {
     int topIndex;
-    double info[MAX_NODES * MAX_ARITY];
+    float info[MAX_NODES * MAX_ARITY];
 } ExStack;
 
 void push(Stack* s, unsigned int info){
@@ -43,14 +43,14 @@ unsigned int pop(Stack* s){
     }
 }
 
-void pushEx(ExStack* s, double info) {
+void pushEx(ExStack* s, float info) {
     (s->topIndex)++;
     if(s->topIndex < MAX_NODES * MAX_ARITY){
         s->info[s->topIndex] = info;
     }
 }
 
-double popEx(ExStack* s) {
+float popEx(ExStack* s) {
     if(s->topIndex >= 0){
         (s->topIndex)--;
         return s->info[(s->topIndex) + 1];
@@ -77,16 +77,16 @@ unsigned int randomFunction(int *seed) {
     return (rand2(seed) % (NUM_FUNCTIONS));
 }
 
-double randomConnectionWeight(int *seed) {
-    return (((double) rand2(seed) / (double) (2147483647) ) * 2 * WEIGTH_RANGE) - WEIGTH_RANGE;
+float randomConnectionWeight(int *seed) {
+    return (((float) rand2(seed) / (float) (2147483647) ) * 2 * WEIGTH_RANGE) - WEIGTH_RANGE;
 }
 
 int randomInterval(int inf_bound, int sup_bound, int *seed) {
     return rand2(seed) % (sup_bound - inf_bound + 1) + inf_bound;
 }
 
-double randomProb(int* seed){
-    return (double)rand2(seed) / 2147483647;//pown(2.0, 31);
+float randomProb(int* seed){
+    return (float)rand2(seed) / 2147483647;//pown(2.0, 31);
 }
 
 
@@ -318,9 +318,9 @@ void mutateTopologyPoint(__global Chromosome *c, __global unsigned int* function
     activateNodes(c);
 }
 
-double executeFunction(__global Chromosome* c, int node, ExStack* exStack){
+float executeFunction(__global Chromosome* c, int node, ExStack* exStack){
     int i;
-    double result, sum;
+    float result, sum;
     unsigned int inputs = c->nodes[node].maxInputs;
     switch (c->nodes[node].function){
         #ifdef ADD
@@ -377,20 +377,20 @@ double executeFunction(__global Chromosome* c, int node, ExStack* exStack){
         
         #ifdef SQ
         case SQ:
-            result = pow((double)popEx(exStack), (double)2);
+            result = pow((float)popEx(exStack), (float)2);
         break;
         #endif
         
         #ifdef CUBE
         case CUBE:
-            result = pow((double)popEx(exStack), (double)3);
+            result = pow((float)popEx(exStack), (float)3);
             break;
         #endif
         
         #ifdef POW
         case POW:
             result = popEx(exStack);
-            result = pow((double)popEx(exStack), (double)result);
+            result = pow((float)popEx(exStack), (float)result);
             break;
         #endif
         
@@ -546,7 +546,7 @@ double executeFunction(__global Chromosome* c, int node, ExStack* exStack){
             for(i = 0; i < inputs; i++){
                 sum += (popEx(exStack) * c->nodes[node].inputsWeight[i]);
             }
-            result = exp(-(pow((double) (sum - 0), (double) 2)) / (2 * pow((double)1, (double)2)));
+            result = exp(-(pow((float) (sum - 0), (float) 2)) / (2 * pow((float)1, (float)2)));
             break;
         #endif
         
@@ -594,9 +594,9 @@ double executeFunction(__global Chromosome* c, int node, ExStack* exStack){
 }
 
 void evaluateCircuitParallel(__global Chromosome* c,
-                            __constant double* data,
-                            __constant double* out,
-                            __local double* error) {
+                            __constant float* data,
+                            __constant float* out,
+                            __local float* error) {
     
     //c->fitness = 0.0;
 
@@ -607,7 +607,7 @@ void evaluateCircuitParallel(__global Chromosome* c,
     int group_id = get_group_id(0);
 
     error[local_id] = 0.0f;
-    double num, div;
+    float num, div;
 
     #ifndef NUM_POINTS_IS_NOT_DIVISIBLE_BY_LOCAL_SIZE
    /* When we know that NUM_POINTS is divisible by LOCAL_SIZE then we can avoid a
@@ -616,18 +616,18 @@ void evaluateCircuitParallel(__global Chromosome* c,
     for(k = 0; k < (M/LOCAL_SIZE) ; k++){
 
     #else
-        for(k = 0; k < ceil( M/ (double)LOCAL_SIZE ) ; k++){
+        for(k = 0; k < ceil( M/ (float)LOCAL_SIZE ) ; k++){
             
             if( k * LOCAL_SIZE + local_id < M){
     #endif
         //printf("c");
         //int i, j;
-        double maxPredicted = -FLT_MAX;
+        float maxPredicted = -FLT_MAX;
         int predictedClass = 0;
         int correctClass = 0;
 
-        double executionOut[MAX_OUTPUTS];
-        double alreadyEvaluated[MAX_NODES];
+        float executionOut[MAX_OUTPUTS];
+        float alreadyEvaluated[MAX_NODES];
         int inputsEvaluatedAux[MAX_NODES];
 
         for(i = 0; i < MAX_NODES; i++){
@@ -717,9 +717,9 @@ void evaluateCircuitParallel(__global Chromosome* c,
 }
 
 void evaluateCircuitParallelValidation(__global Chromosome* c,
-                                    __constant double* data,
-                                    __constant double* out,
-                                    __local double* error) {
+                                    __constant float* data,
+                                    __constant float* out,
+                                    __local float* error) {
     
     //c->fitness = 0.0;
 
@@ -730,7 +730,7 @@ void evaluateCircuitParallelValidation(__global Chromosome* c,
     int group_id = get_group_id(0);
 
     error[local_id] = 0.0f;
-    double num, div;
+    float num, div;
 
     #ifndef NUM_POINTS_VALIDATION_IS_NOT_DIVISIBLE_BY_LOCAL_SIZE_GLOBAL
    /* When we know that NUM_POINTS is divisible by LOCAL_SIZE then we can avoid a
@@ -739,18 +739,18 @@ void evaluateCircuitParallelValidation(__global Chromosome* c,
     for(k = 0; k < (M_VALIDATION/LOCAL_SIZE) ; k++){
 
     #else
-        for(k = 0; k < ceil( M_VALIDATION/ (double)LOCAL_SIZE ) ; k++){
+        for(k = 0; k < ceil( M_VALIDATION/ (float)LOCAL_SIZE ) ; k++){
             
             if( k * LOCAL_SIZE + local_id < M_VALIDATION){
     #endif
         //printf("c");
         //int i, j;
-        double maxPredicted = -FLT_MAX;
+        float maxPredicted = -FLT_MAX;
         int predictedClass = 0;
         int correctClass = 0;
 
-        double executionOut[MAX_OUTPUTS];
-        double alreadyEvaluated[MAX_NODES];
+        float executionOut[MAX_OUTPUTS];
+        float alreadyEvaluated[MAX_NODES];
         int inputsEvaluatedAux[MAX_NODES];
 
         for(i = 0; i < MAX_NODES; i++){
@@ -862,10 +862,10 @@ __kernel void evolve(__global Chromosome* best,
 }
 
 __kernel void evaluate(__global Chromosome* pop,
-                       __constant double* dataset,
-                       __constant double* outputs,
+                       __constant float* dataset,
+                       __constant float* outputs,
                        __global unsigned int* functionSet,
-                       __local double* error){
+                       __local float* error){
 
     int group_id = get_group_id(0);
 
@@ -880,7 +880,7 @@ __kernel void CGP(__global Chromosome* pop,
                   __local int* validIndexes,
                   __global unsigned int* functionSet,
                   __global int *seeds,
-                  __local double* error){
+                  __local float* error){
 
     int group_id = get_group_id(0);
     int seed = seeds[group_id];
